+++ tests/phpunit/Repository/class-votes-repository-test.php
@@ -38,7 +38,7 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 	public function test_create_persists_vote(): void {
 		$repository = new Votes_Repository( $GLOBALS['wpdb'] );
 
-		$id = $repository->create( 1, 'colour', 'John Doe', 42, 9.0 );
+		$id = $repository->create( 1, 'colour', 'John Doe', 42, 9 );
 
 		$this->assertIsInt( $id );
 		$this->assertGreaterThan( 0, $id );
@@ -49,7 +49,7 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 		$this->assertSame( 'colour', $votes[0]->category );
 		$this->assertSame( 'John Doe', $votes[0]->voter_name );
 		$this->assertSame( 42, (int) $votes[0]->image_id );
-		$this->assertSame( 9.0, (float) $votes[0]->score );
+		$this->assertSame( 9, (int) $votes[0]->score );
 	}
 
 	/**
@@ -60,7 +60,7 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 	public function test_create_requires_voter_name(): void {
 		$repository = new Votes_Repository( $GLOBALS['wpdb'] );
 
-		$result = $repository->create( 1, 'colour', '', 42, 9.0 );
+		$result = $repository->create( 1, 'colour', '', 42, 9 );
 
 		$this->assertInstanceOf( \WP_Error::class, $result );
 		$this->assertSame( 'missing_voter_name', $result->get_error_code() );
@@ -74,7 +74,7 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 	public function test_create_validates_score(): void {
 		$repository = new Votes_Repository( $GLOBALS['wpdb'] );
 
-		$result = $repository->create( 1, 'colour', 'John Doe', 42, -1.0 );
+		$result = $repository->create( 1, 'colour', 'John Doe', 42, -1 );
 
 		$this->assertInstanceOf( \WP_Error::class, $result );
 		$this->assertSame( 'invalid_score', $result->get_error_code() );
@@ -88,9 +88,9 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 	public function test_find_by_competition_filters(): void {
 		$repository = new Votes_Repository( $GLOBALS['wpdb'] );
 
-		$repository->create( 1, 'colour', 'Alice', 10, 9.0 );
-		$repository->create( 1, 'bw', 'Alice', 20, 8.0 );
-		$repository->create( 2, 'colour', 'Bob', 30, 7.0 );
+		$repository->create( 1, 'colour', 'Alice', 10, 9 );
+		$repository->create( 1, 'bw', 'Alice', 20, 8 );
+		$repository->create( 2, 'colour', 'Bob', 30, 7 );
 
 		$comp1_all = $repository->find_by_competition( 1 );
 		$this->assertCount( 2, $comp1_all );
@@ -111,9 +111,9 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 	public function test_find_by_image(): void {
 		$repository = new Votes_Repository( $GLOBALS['wpdb'] );
 
-		$repository->create( 1, 'colour', 'Alice', 42, 9.0 );
-		$repository->create( 1, 'colour', 'Bob', 42, 8.0 );
-		$repository->create( 1, 'colour', 'Charlie', 43, 7.0 );
+		$repository->create( 1, 'colour', 'Alice', 42, 9 );
+		$repository->create( 1, 'colour', 'Bob', 42, 8 );
+		$repository->create( 1, 'colour', 'Charlie', 43, 7 );
 
 		$votes = $repository->find_by_image( 42 );
 		$this->assertCount( 2, $votes );
@@ -129,7 +129,7 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 
 		$this->assertFalse( $repository->has_voted( 1, 'colour', 'Alice' ) );
 
-		$repository->create( 1, 'colour', 'Alice', 42, 9.0 );
+		$repository->create( 1, 'colour', 'Alice', 42, 9 );
 
 		$this->assertTrue( $repository->has_voted( 1, 'colour', 'Alice' ) );
 		$this->assertFalse( $repository->has_voted( 1, 'bw', 'Alice' ) );
@@ -144,9 +144,9 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 	public function test_calculate_averages(): void {
 		$repository = new Votes_Repository( $GLOBALS['wpdb'] );
 
-		$repository->create( 1, 'colour', 'Alice', 42, 9.0 );
-		$repository->create( 1, 'colour', 'Bob', 42, 7.0 );
-		$repository->create( 1, 'colour', 'Charlie', 43, 10.0 );
+		$repository->create( 1, 'colour', 'Alice', 42, 9 );
+		$repository->create( 1, 'colour', 'Bob', 42, 7 );
+		$repository->create( 1, 'colour', 'Charlie', 43, 10 );
 
 		$averages = $repository->calculate_averages( 1, 'colour' );
 
@@ -154,11 +154,11 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 		$this->assertArrayHasKey( 43, $averages );
 
 		$this->assertSame( 42, $averages[42]['image_id'] );
-		$this->assertSame( 8.0, $averages[42]['average_score'] );
+		$this->assertSame( 8.0, $averages[42]['average_score'] ); // Average can still be float.
 		$this->assertSame( 2, $averages[42]['vote_count'] );
 
 		$this->assertSame( 43, $averages[43]['image_id'] );
-		$this->assertSame( 10.0, $averages[43]['average_score'] );
+		$this->assertSame( 10.0, $averages[43]['average_score'] ); // Average can still be float.
 		$this->assertSame( 1, $averages[43]['vote_count'] );
 	}
 
@@ -170,9 +170,9 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 	public function test_delete_by_competition(): void {
 		$repository = new Votes_Repository( $GLOBALS['wpdb'] );
 
-		$repository->create( 1, 'colour', 'Alice', 42, 9.0 );
-		$repository->create( 1, 'colour', 'Bob', 43, 8.0 );
-		$repository->create( 2, 'colour', 'Charlie', 44, 7.0 );
+		$repository->create( 1, 'colour', 'Alice', 42, 9 );
+		$repository->create( 1, 'colour', 'Bob', 43, 8 );
+		$repository->create( 2, 'colour', 'Charlie', 44, 7 );
 
 		$this->assertCount( 2, $repository->find_by_competition( 1 ) );
 
@@ -191,10 +191,10 @@ class Votes_Repository_Test extends WP_UnitTestCase {
 	public function test_get_voters(): void {
 		$repository = new Votes_Repository( $GLOBALS['wpdb'] );
 
-		$repository->create( 1, 'colour', 'Alice', 42, 9.0 );
-		$repository->create( 1, 'colour', 'Alice', 43, 8.0 );
-		$repository->create( 1, 'colour', 'Bob', 44, 7.0 );
-		$repository->create( 2, 'colour', 'Charlie', 45, 6.0 );
+		$repository->create( 1, 'colour', 'Alice', 42, 9 );
+		$repository->create( 1, 'colour', 'Alice', 43, 8 );
+		$repository->create( 1, 'colour', 'Bob', 44, 7 );
+		$repository->create( 2, 'colour', 'Charlie', 45, 6 );
 
 		$voters = $repository->get_voters( 1 );
 		$this->assertCount( 2, $voters );
